requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==2.0.1

# Flask Packages
Flask-Login==0.5.0
Flask-Migrate==3.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Flask-User==1.0.2.2
cryptography==3.4.7
gunicorn===20.1.0
Pillow==8.2.0
numpy==1.20.0
requests==2.25.1
# Automated tests
pytest==6.2.4
pytest-cov==2.12.1

